[Code] Implemented getMoves
This is an implementation of getMoves which is going to return a set of
possible moves when a player request for a certain tile. We could do
better that this, by having  an implementation of the boardOutline in
which we are aware of all the kinds of tiles we could place in each
position by mapping the kind of tile needed to the place.